feat: Add support for the yash shell
I wanted to use the `polyglot.sh` script for my current working shell
yash, but the script stated that my shell was not supported. So i
decided to add support for it.

I did my best to adhere to the coding style, and prompt style of the
script. The only thing lacking is the ability to show what Vi mode you
are in since yash does not support that (I looked through its manual
and there is no mention of such a feature at all).

Signed-off-by: Jason Pena <jasonpena@awkless.com>